Add Docker files and scripts for creating AppImage
Added new rules and targets for creating the AppImage because AppImages
need a special handling within a Docker container.

Fixed the CFLAGS usage in CMakeList.txt so that -Wall only applies to
debug builds.  This also fixed an issue where CFLAGS was being replaced
as opposed to appended to, which affected the rules for building an
AppImage.